Added an error in case of missing file during copying.

+++ scripts/build-package.js
@@ -28,6 +28,10 @@ for ( const file of filesToCopy ) {
 	const src = path.join( process.cwd(), file );
 	const dest = path.join( process.cwd(), 'dist', file );
 
+	if ( !fs.existsSync( src ) ) {
+		throw new Error( `File ${ src } does not exist and cannot be copied.` );
+	}
+
 	fs.copyFileSync( src, dest );
 }
